# backend/docs/swagger.yaml
basePath: /
definitions:
...
  admin.MessageResponse:
    properties:
      message:
        type: string
    type: object
  admin.ModuleDetailResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      download_count:
        type: integer
      id:
        type: string
      name:
        type: string
      namespace:
        type: string
      source:
        type: string
      system:
        type: string
      updated_at:
        type: string
      versions:
        items:
          $ref: '#/definitions/admin.ModuleVersionItem'
        type: array
    type: object
...
  admin.ModuleVersionItem:
    properties:
      created_at:
        type: string
      deprecated:
        type: boolean
      deprecated_at: {}
      deprecation_message: {}
      download_count:
        type: integer
      id:
        type: string
      version:
        type: string
    type: object
...
  mirror.MirrorArchiveEntry:
    properties:
      hashes:
        additionalProperties:
          type: string
        type: object
      url:
        type: string
    type: object
  mirror.MirrorPlatformIndexResponse:
    properties:
      archives:
        additionalProperties:
          $ref: '#/definitions/mirror.MirrorArchiveEntry'
        type: object
    type: object
...
  modules.ModuleUploadResponse:
    properties:
      checksum:
        type: string
      created_at:
        type: string
      filename:
        type: string
      id:
        type: string
      name:
        type: string
      namespace:
        type: string
      size_bytes:
        type: integer
      system:
        type: string
      version:
        type: string
    type: object
...
info:
  contact:
    email: admin@your-registry.example.com
    name: Registry Support
  description: Complete Terraform Module and Provider Registry with SCM integration,
    mirrors, and storage backends
  license:
    name: Apache-2.0
  termsOfService: https://your-registry.example.com/terms
  title: Terraform Registry API
  version: 1.0.0
paths:
...
  /api/v1/modules:
    post:
      consumes:
      - multipart/form-data
      description: Upload a new module version as a .tar.gz archive. Creates the module
        if it doesn't exist. Requires modules:publish scope.
      parameters:
      - description: Module namespace
        in: formData
        name: namespace
        required: true
        type: string
      - description: Module name
        in: formData
        name: name
        required: true
        type: string
      - description: Target system (e.g. aws, azurerm)
        in: formData
        name: system
        required: true
        type: string
      - description: Semantic version (e.g. 1.2.3)
        in: formData
        name: version
        required: true
        type: string
      - description: Module description
        in: formData
        name: description
        type: string
      - description: Source URL
        in: formData
        name: source
        type: string
      - description: Module archive (.tar.gz, max 100MB)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/modules.ModuleUploadResponse'
        "400":
          description: Invalid request, bad version format, or invalid archive
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Version already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Upload module version
      tags:
      - Modules
  /api/v1/modules/{namespace}/{name}/{system}:
    delete:
      description: Delete a module and all its versions, including files in storage.
        Requires modules:delete scope.
      parameters:
      - description: Module namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Module name
        in: path
        name: name
        required: true
        type: string
      - description: Target system (e.g. aws, azurerm)
        in: path
        name: system
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Module not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete module
      tags:
      - Modules
    get:
      description: Retrieve a module with all its versions, download counts, and metadata.
        No authentication required; authentication is optional and provides user context.
      parameters:
      - description: Module namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Module name
        in: path
        name: name
        required: true
        type: string
      - description: Target system (e.g. aws, azurerm)
        in: path
        name: system
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.ModuleDetailResponse'
        "404":
          description: Module not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get module
      tags:
      - Modules
...
  /terraform/providers/{hostname}/{namespace}/{type}/{versionfile}:
    get:
      description: Returns download URLs and hashes for all platforms of a specific
        provider version, per the Terraform Network Mirror Protocol.
      parameters:
      - description: Origin registry hostname (e.g. registry.terraform.io)
        in: path
        name: hostname
        required: true
        type: string
      - description: Provider namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Provider type (e.g. aws, azurerm)
        in: path
        name: type
        required: true
        type: string
      - description: Version with .json suffix (e.g. 1.2.3.json)
        in: path
        name: versionfile
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mirror.MirrorPlatformIndexResponse'
        "400":
          description: Invalid version format
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Provider or version not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Network mirror provider platform index
      tags:
      - Mirror Protocol
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: '"JWT token or API key. For JWT: ''Bearer {token}''. For API Key:
      ''Bearer {api_key}''"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"